+++ src/components/send-payment/send-amount.tsx
@@ -1,7 +1,7 @@
 import React, { ChangeEvent } from "react";
 import BigNumber from "bignumber.js";
 import { Button, Heading, Input } from "@stellar/design-system";
-import { formatTokenAmount } from "utils/format";
+import { formatTokenAmount } from "helpers/format";
 
 interface SendAmountProps {
   amount: string;
